refactor: move autocomplete outside of ui
